Move LocationPredictor and related classes to separate module
